dev_projetos/flaskapp-metrics/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-prometheus==0.46b0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.1
prometheus_client==0.20.0
setuptools==70.3.0
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
